scripts: dtc: Remove gpio binding properties check
This change appeared here as a result of the import of a DTC snapshot
from 5.10 upstream kernel in commit da03aca ("scrits: dtc: Import from
Linux5.10"). Since in the current kernel version the -gpio/-gpios suffix
is widely used to describe the number of GPIOs supported by the
controller, revert this change in order not to break all current SoCs.

This should be applicable as of 5.4 kernel.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>