fscrypt: add missing fscrypt_symlink_getattr
This was missed during a previous merge which wrongly assumed that
fscrypt_supp.h was missing from downstream. In reality, the file is
simply named fscrypt.h instead in android-common kernels and these
patches shall not be skipped.

This function definition is required for the 4.19.207 merge. Previously,
other changes in relation to fscrypt have been skipped for the same
reason, and we should probably port them back at some point. One example
is `fscrypt_is_nokey_name` from 4.19.165.